requirements.txt
Django~=5.1.6
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
pre-commit~=4.2.0
pyright[nodejs]~=1.1.398
django-stubs-ext~=5.1.3
django-stubs~=5.1.3
djangorestframework~=3.16.0
djangorestframework-dataclasses~=1.3.1
drf-yasg~=1.21.10
djangorestframework-stubs~=3.15.3
drf-yasg-stubs~=0.1.4
celery~=5.5.1
django-celery-results~=2.6.0
django-celery-beat~=2.7.0
celery-types~=0.23.0
pdf2image~=1.17.0